ceph: add namespace to teardown in finalizer update
Add the `-n rook-ceph` namespace to the `kubectl patch` as part of the
finalizer update in teardown troubleshooting.

I had to add the namespace identifier, as all the "gets" had the `rook-ceph`
namespace, but the patch did not. If I had a leftover "replicapool"
it was unable to patch the replicapool without the namespace.

Also, adjust the list of CRDs affected by the finalizer update for
version 1.5.

Signed-off-by: Stephen Rust <srust@srust.dev>